+++ src/main/java/com/konstl/dormitories/agreement/AgreementService.java
@@ -4,9 +4,8 @@ import com.konstl.dormitories.agreement.dto.AgreementResponse;
 import com.konstl.dormitories.agreement.dto.CreateAgreementRequest;
 import com.konstl.dormitories.agreement.dto.UpdateAgreementRequest;
 import com.konstl.dormitories.utils.PageResponse;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
@@ -20,7 +19,7 @@ public interface AgreementService {
 
     PageResponse<AgreementResponse> findByEvictionDate(LocalDate date, int page, int size);
 
-    PageResponse<AgreementResponse> findByCost(Double cost, int page, int size);
+    PageResponse<AgreementResponse> findByCost(BigDecimal cost, int page, int size);
 
     PageResponse<AgreementResponse> findByMoveInDateGreaterThanEqual(LocalDate date, int page, int size);
 
@@ -34,11 +33,11 @@ public interface AgreementService {
 
     PageResponse<AgreementResponse> findByEvictionDateBetween(LocalDate startDate, LocalDate endDate, int page, int size);
 
-    PageResponse<AgreementResponse> findByCostBetween(Double startCost, Double endCost, int page, int size);
+    PageResponse<AgreementResponse> findByCostBetween(BigDecimal startCost, BigDecimal endCost, int page, int size);
 
-    PageResponse<AgreementResponse> findByCostGreaterThanEqual(Double cost, int page, int size);
+    PageResponse<AgreementResponse> findByCostGreaterThanEqual(BigDecimal cost, int page, int size);
 
-    PageResponse<AgreementResponse> findByCostLessThanEqual(Double cost, int page, int size);
+    PageResponse<AgreementResponse> findByCostLessThanEqual(BigDecimal cost, int page, int size);
 
     AgreementResponse create(CreateAgreementRequest createRequest);
