+++ src/main/java/com/konstl/dormitories/position/dto/CreatePositionRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreatePositionRequest {
 
@@ -13,5 +15,5 @@ public class CreatePositionRequest {
     private String name;
 
     @NotNull
-    private Double salary;
+    private BigDecimal salary;
 }
